Rename share_request_dialog module to share_request
