+++ src/widgets/mod.rs
@@ -1,7 +1,7 @@
 mod file_card;
 mod recipient_card;
-mod share_request_dialog;
+mod share_request;
 
 pub use file_card::*;
 pub use recipient_card::*;
-pub use share_request_dialog::*;
+pub use share_request::*;
